Update the server API and drawData.
This includes:
Converting the uri into AbstractSQL, adding support for Insert/Update/Delete queries in the AbstractSQL/SQL generation.
Using an SBVR model for storing the UI textarea data.
Using the new AsyncQueueCallback in drawData for the callbacks.
Using a navigable tree object in drawData to simplify tree traversal.
Fixing the double drawData and other weirdness on page reloads.
JSON stringifying and parsing in the express emulator to act more like a real server.