Add meta endpoint `$canAccess` for every api root, to allow checking resource access for arbitrary actions
This adds an `$canAccess` endpoint to every api root. This endpoint accepts POST requests
with parameters `resource`, `action` and `id`.
- The `resource` parameter holds the resource name in question.
- The `action` parameter holds the action in question.
- The `id` parameter holds the resource id in question.

The endpoint responds with an empty response with http status code 200 if access is granted.
The endpoint responds with an http status code 401 if access is not granted.
The endpoint can also respond with http status code 400 or 500 in case of client or server errors.

Change-Type: minor
Connects-To: #91
Signed-off-by: Andreas Fitzek <andreas@resin.io>