Update `@balena/sbvr-types` to enable ISODatestring as internal pine API interface for date and date time fields.
Update `@balena/sbvr-types` from 6.1.1 to 7.0.1

pinejs provides an internal and external interface. The external interface returns ISODateString encoded data.
The internal interface should be consistent to the external interface, so that the pinejs api internally used also returns an ISODateString encoded data. This is solved in `sbvr-types` and is a major in the submodule and made this also a major change for pinejs itself.

https://github.com/balena-io-modules/sbvr-types/pull/91

Change-type: major
Signed-off-by: Harald Fischer <harald@balena.io>