Replace the cached abstractSqlQuery with the compiled sqlQuery once generated
This allows freeing up all the memory used by the abstractSqlQuery tree
once the sqlQuery has been generated at the cost that the
abstractSqlQuery cannot be recompiled for another db engines. In
practice this shouldn't be a limitation as using multiple db engines
with a single model is not supported and unlikely to work.

Change-type: major