Improve webresource tests to be independent of filename
Previous version of tests were relying on the fact that the href (generated) property used the given filename to be generated and would search over the bucket for that filename.

This is highly undesirable as it makes the tests unreliable to possible underlying (non breaking) changes, as it relies on a side effect

The current PR changes it by ensuring the specific keys exist on the bucket without any relation for the given filename.

See: https://balena.fibery.io/search/tyKzx#Work/Improvement/Make-webresources-file-deletion-more-reliable-2819
Change-type: patch